chatbot_project/requirements.txt
Django>=4.2
channels>=4.0
channels-redis>=4.0   # opcional (recomendado para producción / múltiples procesos)
groq
python-dotenv
